Updated ncbi api processing to use multiprocessing instead of threading Split apiworker into a separate ncbi library module to allow functionality with importlib Updated processing library to only remove added system paths after module execution
